\name{tabus-package}
\alias{tabus-package}
\alias{tabus}
\docType{package}
\title{
\packageTitle{tabus}
}
\description{
\packageDescription{tabus}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{tabus}
\packageIndices{tabus}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{tabus}

Maintainer: \packageMaintainer{tabus}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
